requirements.txt
numpy~=1.22.2
pandas~=1.4.0
matplotlib~=3.5.1
seaborn~=0.11.2
biopython~=1.79
tqdm~=4.62.3
PyYAML~=6.0
